Added more functionality to the BuyerPage
Created the search functionality via text AND categories.

Added in a CurrentUser class for the addToCart feature, needed to have a
way to link the UserId to the bookID

Added in a Cart table in the database on DatabaseSetup.java. If you need
to make changes to the database. manually run DatabaseSetup.java then
run main.java.

LoginPage adjusted to match Update CurrentUser when logging in for the
cart.

Added a remove duplicates class to be manually ran. Accidently kept
adding extra test data and needed a way to easily remove duplicates.

Added a databaseViewer class to make sure the view in the application
matched what was in the actual database. This needs to be ran manually
as well.

//------ TODO ------\\
* Make the categories load dynamically
* Create the cart page so we can view it
* On the cart page, make it so user can remove items from the cart.