Updating database logic for Cart+Seller
Complete rehaul of the Seller and Cart page to match logic for the
database. Did a bit of copy and paste between pages. Kept all formatting
from Tsvitanasuk, changed the grid layout to mimic the book pull from
the Buyer page on CartPage. A lot more changes were made to the cart
page due to setup. Seller page only added in the logic for the listBook
Function, and wrapped the mainUI in a scrollPane for usablility.